fixes ui issues with setup, gameboard, and pieces
